Update Solve The Maze.cpp

+++ 1700/Solve The Maze.cpp
@@ -18,7 +18,7 @@ int main(){
 	for(cin>>t1;t1--;)
         {
 		cin>>n>>m;
-		for(int i=1;i<=n;i++)scanf("%s",a[i]+1);
+		for(ll i=1;i<=n;i++)scanf("%s",a[i]+1);
 		for(int i=0;i<=n+1;i++) for(int j=0;j<=m+1;j++) b[i][j]=v[i][j]=0;
 		for(int i=1;i<=n;i++) for(int j=1;j<=m;j++){
 		   if(a[i][j]=='#') b[i][j]=1;
